+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 ### 0.2.8 (2021-02-17)
 
 * Change logger to use `Level` rather than `quiet` - add `Logger.apply(quiet, ...)` for backward compatibility
+* Remove `logger` parameter from `SysUtils.execCommand` and instead use custom Exception classes
 
 ### 0.2.7 (2021-02-17)